feat: create CategoryCard entity
