+++ src/entities/CategoryCard/index.ts
@@ -1,0 +1,3 @@
+import CategoryCard from './ui/CategoryCard'
+
+export { CategoryCard }
