+++ src/entities/CategoryCard/ui/CategoryCard.module.scss
@@ -1,0 +1,23 @@
+.image {
+  width: 100%;
+  height: auto;
+  border-radius: 6px;
+  aspect-ratio: 1/1;
+  object-fit: cover;
+}
+
+.subtitle {
+  font: var(--font-m);
+  margin: 0;
+  color: var(--text-primary);
+}
+
+.text {
+  margin: 0;
+  font: var(--font-m);
+  color: var(--text-secondary);
+
+  &:last-child {
+    margin-top: auto;
+  }
+}
