+++ src/pages/Category/ui/Category.module.scss
@@ -45,30 +45,6 @@
         transform: translateY(-5px);
         box-shadow: 0 6px 12px var(--shadow-hover);
       }
-
-      .image {
-        width: 100%;
-        height: auto;
-        border-radius: 6px;
-        aspect-ratio: 1/1;
-        object-fit: cover;
-      }
-
-      .subtitle {
-        font: var(--font-m);
-        margin: 0;
-        color: var(--text-primary);
-      }
-
-      .text {
-        margin: 0;
-        font: var(--font-m);
-        color: var(--text-secondary);
-
-        &:last-child {
-          margin-top: auto;
-        }
-      }
     }
   }
 }
